+++ Engine2/src/resources/shaders/lines.geom
@@ -4,7 +4,7 @@ layout(lines_adjacency) in;
 layout(triangle_strip, max_vertices = 7) out;
 
 uniform mat4 pv;
-uniform vec2 viewport;
+uniform ivec2 viewport;
 uniform float thickness;
 
 vec3 toScreenSpace(vec4 v)
