+++ Engine2/src/resources/shaders/point.geom
@@ -6,7 +6,7 @@ layout(points) in;
 layout(triangle_strip, max_vertices = 16) out;
 
 uniform float thickness;
-uniform vec2 viewport;
+uniform ivec2 viewport;
 
 vec3 toScreenSpace(vec4 v)
 {
